News Classification Using SpaCy
This project focuses on developing a robust fake news detection model using a combination of SpaCy for natural language processing, and machine learning algorithms such as Multinomial Naive Bayes (MultinomialNB) and K-Neighbors Classifier. The goal is to accurately classify news articles as fake or real based on their textual content.